Method: Simplify resolveTemplateType()
The previous code tried to resolve template types using methods
found on parent classes and interfaces of the class we're examining.
This doesn't work reliably, because they can take completely different
template parameters (and we don't support generic interfaces anyway).

It shouldn't be needed in the first place: much like in 0640175de8,
the actual resolution of inherited template parameter names happens
in Clazz when copying the inherited method.