Clazz: Move Method/Property cloning code to their classes
Method already had a similar method, used when cloning method for a
one-time use of a generic method with known types.

The code for Property had a try...catch around it, but IssueException
can never be thrown there.

Somehow this causes Phan to detect a new issue with documentation in
its own code. No idea how that happened.