Handle PhanAttributeWrongTarget for constructor property promotion
For *internal* attributes, if used on a promoted property and the attribute
targets properties or parameters but not both, it gets recorded on the element
it supports and ignored on the element that it does not support. The only
existing such attribute is `#[\SensitiveParameter]` which supports parameters
but not properties; to verify that internal attributes that support properties
but not parameters are also handled properly, a fake internal attribute
`#[\InternalAttribForProperties]` is used.