Improve handling of promoted properties also declared via doc comments
When a property is declared via PHP doc comments, `ParseVisitor::addProperty()`
will return null, previously causing
`ParseVisitor::addPromotedConstructorPropertyFromParam()` to skip the remainder
of its handling. This led to
- false positives for `PhanReadOnlyPHPDocProperty`
- false negatives for `PhanAttributeWrongTarget`

Now, if the `addProperty()` call fails,
`addPromotedConstructorPropertyFromParam()` will check for an existing property
declared with doc comments and, if one exists, update it the same as it would
update a property not declared with doc comments.

Fixes #5022