Improve handling of never-typed functions and methods
When a function or method is typed to not return anything, if the function
might not return or throw, instead of complain that the function might "fail to
return a value", complain that the function might "try to return a value",
because the expectation is reversed. In cases where the function explicitly
does return, `PhanSyntaxReturnStatementInNever` will complain.

The new error messages are from new issues `PhanPluginNeverReturnMethod` and
`PhanPluginAlwaysReturnMethod`.

If a function or method calls another function or method, try to infer of that
other method will always throw; global functions, static methods, and instance
methods on `$this` should all be handled.

Fixes #4779