Normalize generic multi-type elements before expanding them
Phan only supports union types inside generic type parameters
by expanding them to an union of simple generic types, that is,
`(A|null)[]` becomes `A[]|null[]`. The temporary value that
holds a union inside the type param is called a multi-type.

As a result, while `A|null` and `?A` are treated as the same type,
`(A|null)[]` and `(?A)[]` were not. It seems difficult to fix this
in the type system, it would require "deeper" normalization than we
currently do. Instead, normalize before expanding the multi-types,
so that `(A|null)[]` becomes `(?A)[]` instead of `A[]|null[]`.
This avoids problems at least in this common case. Fixes #5049.