Allow `@extends`/`@inherits` in PHPDoc on traits again
`@extends`/`@inherits` can be used to document that the trait
is only used on that class, and avoid issues about undeclared
methods/properties/constants when they're defined in that class.

I'm not sure if this is supposed to work, but it does, see #5002.
I unintentionally broke it in 1193d813f2. Add some tests so that
it doesn't get accidentally broken again.