Bring back previously deleted test files
These tests required the specified PHP version //as minimum//, so they
should still be run now that we require PHP 8.1+, but they can be added
to the default suite rather than having bespoke test classes. It's
possible that we'll be overtesting something, but that should be better
than undertesting. For now, tests are restored to their original
location. The next commit will move them.