Make PHPUnit suites not overlap
Internall, PHPUnit assumes that test suites are non-overlapping. Not
doing that is harmless in PHPUnit 9, but causes issues in PHPUnit 10
(https://github.com/sebastianbergmann/phpunit/issues/5670), and will
throw an error in PHPUnit 11.

So, drop the catch-all "All" suite, and add the missing tests to other
suite. A new "Misc" suite is introduced for single-purpose things that
don't seem to fit elsewhere, although in future it would be nice to
restructure this.