Fix CI
Including reverts of a couple of careless changes I made to test
expectations in 8dd67bd.

Also remove an outdated error filter for PHPUnit 8.

Make a test not use PHP short tags, as those may or may not work in
different environments (depending on the short_open_tag ini setting).