Infer truthiness of func_get_args and array elements
For func_get_args, check if the current function has at least one
required parameter, and if so infer that the return value of
func_get_args is a non-empty list. Else just stick with list as we knew
from reflection.

In ConditionVisitor, implement `visitDim` so that we can apply type
modifications to individual array elements.